Implement read-write locks and simple testing.
Implement read-write locks. Have macros defined to print status of data
structures in read-write lock objects or assert valid state if and only
if the DEBUG flag is set by the compiler in the Makefile.

Implement a small testing infrastructure that allows each individual
test to be defined in a separate classes and reports how long each test
takes in microseconds. Results are simply printed to stdout.

Create three small tests that in total test all of the read-write lock
functions in single threads.

Clarify in readme that read-write locks are designed to be
writer-biased.