Adjust implementation and tests, and add a test:
Expand the readme to explain build and run process.

Change the rwlock internals to have pointers in the struct, as the
existence of the `rwlock_uninit` function fits the use case of an API
that expects data in the rwlock object to be manually deallocated.

Add one test with two threads. Adjust failure reporting to have
plurals.