Fix a correctness bug in rwlock implementation.
Have the first active reader in a series of concurrent readers
effectively acquire the write lock in order to prevent writers from
starting to write while readers that had previously started to read
continue to read.

Clearly outline requirements for every mechanism in each rwlock function
in order to account for every line of code and for potential instructive
purposes.